pyproject.toml
[tool.poetry]
name = "bootcamp-aberto-aovivo"
version = "0.1.0"
description = ""
authors = ["lvgalvaofilho <lvgalvaofilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.11.5"
apache-airflow = "^2.9.0"
apache-airflow-providers-dbt-cloud = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
